+++ src/Audit.EntityFramework/ConfigurationApi/AuditEntityMapping.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Audit.Core;
 
 namespace Audit.EntityFramework.ConfigurationApi
@@ -8,8 +9,8 @@ namespace Audit.EntityFramework.ConfigurationApi
     {
         private Dictionary<Type, MappingInfo> _mapping = new Dictionary<Type, MappingInfo>();
         private List<KeyValuePair<Func<EventEntry, bool>, MappingInfo>> _explicitMapping = new List<KeyValuePair<Func<EventEntry, bool>, MappingInfo>>();
-        private Func<AuditEvent, EventEntry, object, bool> _commonAction = null;
-        
+        private Func<AuditEvent, EventEntry, object, Task<bool>> _commonAction = null;
+
 
         public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>()
         {
@@ -25,6 +26,20 @@ namespace Audit.EntityFramework.ConfigurationApi
                 Action = (ev, ent, auditEntity) =>
                 {
                     entityAction.Invoke(ev, ent, (TAuditEntity)auditEntity);
+                    return Task.FromResult(true);
+                }
+            };
+            return this;
+        }
+
+        public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>(Func<AuditEvent, EventEntry, TAuditEntity, Task> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, ent, auditEntity) =>
+                {
+                    await entityAsyncAction.Invoke(ev, ent, (TAuditEntity)auditEntity);
                     return true;
                 }
             };
@@ -36,7 +51,17 @@ namespace Audit.EntityFramework.ConfigurationApi
             _mapping[typeof(TSourceEntity)] = new MappingInfo()
             {
                 TargetType = typeof(TAuditEntity),
-                Action = (ev, ent, auditEntity) => entityAction.Invoke(ev, ent, (TAuditEntity)auditEntity)
+                Action = (ev, ent, auditEntity) => Task.FromResult(entityAction.Invoke(ev, ent, (TAuditEntity)auditEntity))
+            };
+            return this;
+        }
+
+        public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>(Func<AuditEvent, EventEntry, TAuditEntity, Task<bool>> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, ent, auditEntity) => await entityAsyncAction.Invoke(ev, ent, (TAuditEntity)auditEntity)
             };
             return this;
         }
@@ -49,18 +74,44 @@ namespace Audit.EntityFramework.ConfigurationApi
                 Action = (ev, ent, auditEntity) =>
                 {
                     entityAction.Invoke((TSourceEntity)ent.Entry.Entity, (TAuditEntity)auditEntity);
+                    return Task.FromResult(true);
+                }
+            };
+            return this;
+        }
+
+        public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>(Func<TSourceEntity, TAuditEntity, Task> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, ent, auditEntity) =>
+                {
+                    await entityAsyncAction.Invoke((TSourceEntity)ent.Entry.Entity, (TAuditEntity)auditEntity);
                     return true;
                 }
             };
             return this;
         }
 
+
         public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>(Func<TSourceEntity, TAuditEntity, bool> entityAction)
         {
             _mapping[typeof(TSourceEntity)] = new MappingInfo()
             {
                 TargetType = typeof(TAuditEntity),
-                Action = (ev, ent, auditEntity) => entityAction.Invoke((TSourceEntity)ent.Entry.Entity, (TAuditEntity)auditEntity)
+                Action = (ev, ent, auditEntity) => Task.FromResult(entityAction.Invoke((TSourceEntity)ent.Entry.Entity, (TAuditEntity)auditEntity))
+            };
+            return this;
+        }
+
+
+        public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>(Func<TSourceEntity, TAuditEntity, Task<bool>> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, ent, auditEntity) => await entityAsyncAction.Invoke((TSourceEntity)ent.Entry.Entity, (TAuditEntity)auditEntity)
             };
             return this;
         }
@@ -70,7 +121,17 @@ namespace Audit.EntityFramework.ConfigurationApi
             _mapping[typeof(TSourceEntity)] = new MappingInfo()
             {
                 TargetTypeMapper = mapper,
-                Action = entityAction
+                Action = (ev, ent, auditEntity) => Task.FromResult(entityAction.Invoke(ev, ent, auditEntity))
+            };
+            return this;
+        }
+
+        public IAuditEntityMapping Map<TSourceEntity>(Func<EventEntry, Type> mapper, Func<AuditEvent, EventEntry, object, Task<bool>> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetTypeMapper = mapper,
+                Action = async (ev, ent, auditEntity) => await entityAsyncAction.Invoke(ev, ent, auditEntity)
             };
             return this;
         }
@@ -83,6 +144,20 @@ namespace Audit.EntityFramework.ConfigurationApi
                 Action = (ev, entry, entity) =>
                 {
                     entityAction.Invoke(ev, entry, entity);
+                    return Task.FromResult(true);
+                }
+            };
+            return this;
+        }
+
+        public IAuditEntityMapping Map<TSourceEntity>(Func<EventEntry, Type> mapper, Func<AuditEvent, EventEntry, object, Task> entityAsyncAction)
+        {
+            _mapping[typeof(TSourceEntity)] = new MappingInfo()
+            {
+                TargetTypeMapper = mapper,
+                Action = async (ev, entry, entity) =>
+                {
+                    await entityAsyncAction.Invoke(ev, entry, entity);
                     return true;
                 }
             };
@@ -106,6 +181,20 @@ namespace Audit.EntityFramework.ConfigurationApi
                 Action = (ev, entry, entity) =>
                 {
                     entityAction?.Invoke(entry, (TAuditEntity)entity);
+                    return Task.FromResult(true);
+                }
+            }));
+            return this;
+        }
+
+        public IAuditEntityMapping MapExplicit<TAuditEntity>(Func<EventEntry, bool> predicate, Func<EventEntry, TAuditEntity, Task> entityAsyncAction = null)
+        {
+            _explicitMapping.Add(new KeyValuePair<Func<EventEntry, bool>, MappingInfo>(predicate, new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, entry, entity) =>
+                {
+                    await entityAsyncAction?.Invoke(entry, (TAuditEntity)entity);
                     return true;
                 }
             }));
@@ -120,6 +209,20 @@ namespace Audit.EntityFramework.ConfigurationApi
                 Action = (ev, entry, entity) =>
                 {
                     entityAction?.Invoke(entry, (TAuditEntity)entity);
+                    return Task.FromResult(true);
+                }
+            }));
+            return this;
+        }
+
+        public IAuditEntityMapping MapTable<TAuditEntity>(string tableName, Func<EventEntry, TAuditEntity, Task> entityAsyncAction = null)
+        {
+            _explicitMapping.Add(new KeyValuePair<Func<EventEntry, bool>, MappingInfo>(ent => ent.Table?.Equals(tableName) == true, new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = async (ev, entry, entity) =>
+                {
+                    await entityAsyncAction?.Invoke(entry, (TAuditEntity)entity);
                     return true;
                 }
             }));
@@ -131,13 +234,13 @@ namespace Audit.EntityFramework.ConfigurationApi
             _commonAction = (ev, ent, auditEntity) =>
             {
                 entityAction.Invoke(ev, ent, auditEntity);
-                return true;
+                return Task.FromResult(true);
             };
         }
 
         public void AuditEntityAction(Func<AuditEvent, EventEntry, object, bool> entityAction)
         {
-            _commonAction = (ev, ent, auditEntity) => entityAction.Invoke(ev, ent, auditEntity);
+            _commonAction = (ev, ent, auditEntity) => Task.FromResult(entityAction.Invoke(ev, ent, auditEntity));
         }
 
         public void AuditEntityAction<T>(Action<AuditEvent, EventEntry, T> entityAction)
@@ -145,13 +248,41 @@ namespace Audit.EntityFramework.ConfigurationApi
             _commonAction = (ev, ent, auditEntity) =>
             {
                 entityAction.Invoke(ev, ent, (T)auditEntity);
-                return true;
+                return Task.FromResult(true);
             };
         }
 
         public void AuditEntityAction<T>(Func<AuditEvent, EventEntry, T, bool> entityAction)
         {
-            _commonAction = (ev, ent, auditEntity) => entityAction.Invoke(ev, ent, (T)auditEntity);
+            _commonAction = (ev, ent, auditEntity) => Task.FromResult(entityAction.Invoke(ev, ent, (T)auditEntity));
+        }
+
+        public void AuditEntityAction(Func<AuditEvent, EventEntry, object, Task<bool>> entityAsyncAction)
+        {
+            _commonAction = async (ev, ent, auditEntity) => await entityAsyncAction.Invoke(ev, ent, auditEntity);
+        }
+
+        public void AuditEntityAction<T>(Func<AuditEvent, EventEntry, T, Task<bool>> entityAsyncAction)
+        {
+            _commonAction = async (ev, ent, auditEntity) => await entityAsyncAction.Invoke(ev, ent, (T)auditEntity);
+        }
+
+        public void AuditEntityAction(Func<AuditEvent, EventEntry, object, Task> entityAsyncAction)
+        {
+            _commonAction = async (ev, ent, auditEntity) =>
+            {
+                await entityAsyncAction.Invoke(ev, ent, auditEntity);
+                return true;
+            };
+        }
+
+        public void AuditEntityAction<T>(Func<AuditEvent, EventEntry, T, Task> entityAsyncAction)
+        {
+            _commonAction = async (ev, ent, auditEntity) =>
+            {
+                await entityAsyncAction.Invoke(ev, ent, (T)auditEntity);
+                return true;
+            };
         }
 
         internal Func<Type, EventEntry, Type> GetMapper()
@@ -171,7 +302,7 @@ namespace Audit.EntityFramework.ConfigurationApi
                 {
                     return null;
                 }
-                foreach(var em in _explicitMapping)
+                foreach (var em in _explicitMapping)
                 {
                     if (em.Key.Invoke(e))
                     {
@@ -183,9 +314,9 @@ namespace Audit.EntityFramework.ConfigurationApi
         }
 
         // Returns a generic action that executes the specific action for the mapping and the common action
-        internal Func<AuditEvent, EventEntry, object, bool> GetAction()
+        internal Func<AuditEvent, EventEntry, object, Task<bool>> GetAction()
         {
-            return (ev, ent, auditEntity) =>
+            return async (ev, ent, auditEntity) =>
             {
                 MappingInfo map = null;
                 bool include = true;
@@ -197,7 +328,7 @@ namespace Audit.EntityFramework.ConfigurationApi
                         if (em.Key.Invoke(ent))
                         {
                             map = em.Value;
-                            include = map.Action.Invoke(ev, ent, auditEntity);
+                            include = await map.Action.Invoke(ev, ent, auditEntity);
                             mappedExplicitly = true;
                             break;
                         }
@@ -210,18 +341,17 @@ namespace Audit.EntityFramework.ConfigurationApi
                     {
                         if (map.Action != null)
                         {
-                            include = map.Action.Invoke(ev, ent, auditEntity);
+                            include = await map.Action.Invoke(ev, ent, auditEntity);
                         }
                     }
                 }
+              
                 if (include && _commonAction != null)
                 {
-                    include = _commonAction.Invoke(ev, ent, auditEntity);
+                    include = await _commonAction.Invoke(ev, ent, auditEntity);
                 }
                 return include;
             };
         }
-
-
     }
 }