+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [17.0.3] - 2021-??-??
 ### Modified
 - Audit.NET.MySql: Fix null reference exception for mySQL data provider when using custom transactions (#395)
+- Audit.NET.EntityFramework / Audit.NET.EntityFramework.Core: Adding Support for asynchronous AuditEntityAction for the Entity Framework Data Provider (#397)
 
 ## [17.0.2] - 2021-04-21
 ### Modified